+++ examples/09_publish_loop_forever.jl
@@ -22,5 +22,3 @@ topic = "test/julia"
 end
 
 loop_forever(client)
-
-