+++ src/Mosquitto.jl
@@ -3,7 +3,7 @@
 # https://github.com/eclipse/mosquitto/blob/master/include/mosquitto.h
 module Mosquitto
 
-import Base: finalizer, iterate
+import Base: finalizer, iterate, show
 using Random
 
 include("MosquittoCwrapper/MosquittoCwrapper.jl")
@@ -29,7 +29,7 @@ export tls_set, tls_psk_set
 export will_set, will_clear
 
 include("properties_v5.jl")
-export PropertyList, create_property_list, add_property!, read_property_list
+export PropertyList, Property, create_property_list, add_property!, read_property_list
 
 include("callbacks_v5.jl")
 include("client_v5.jl")