+++ src/properties_v5.jl
@@ -9,8 +9,8 @@ mutable struct PropertyList
     mosq_prop::Base.RefValue{Ptr{CmosquittoProperty}}
 
     function PropertyList()
-        proplist = new( Ref( Ptr{CmosquittoProperty}(UInt(C_NULL))) )
-        
+        proplist = new(Ref(Ptr{CmosquittoProperty}(UInt(C_NULL))))
+
         finalizer(proplist) do x
             MosquittoCwrapper.property_free_all(x.mosq_prop)
             return nothing
@@ -26,14 +26,17 @@ function PropertyList(name::String, value::T) where {T}
     return proplist
 end
 
+
 """
 struct Property
 
 A struct with fields:
-* name:: String -> the topic where the message was received from
+* name:: String -> the name of the property, e.g., "Content-Type", string representation of the `mqtt5_property` (except for user-property, for which this is the user provided key instead).
 * value::Vector{UInt8} -> the value as a byte array
 * prop::MosquittoCwrapper.mqtt5_property -> the assigned MQTT5 property
 * type::MosquittoCwrapper.mqtt5_property_type -> the data type that comes with this MQTT5 property
+
+This type is returned from `read_property_list(::PropertyList)` or can be constructed using `Property(name::String, value)`.
 """
 struct Property
     name::String
@@ -45,12 +48,12 @@ struct Property
         # get first identifier if any
         prop_id, success = MosquittoCwrapper.property_identifier(prop_ptr)
         !success && error("Could not create Property from pointer $prop_ptr")
-    
+
         # get name and type as well
         name = MosquittoCwrapper.property_identifier_to_string(prop_id)
         msg, _, type = MosquittoCwrapper.string_to_property_info(name)
         msg != MosquittoCwrapper.MOSQ_ERR_SUCCESS && error("Couldnt find property info for $name")
-    
+
         # get value as byte array
         if type == MosquittoCwrapper.MQTT_PROP_TYPE_STRING_PAIR
             pair, success = MosquittoCwrapper.property_read_string_pair(prop_ptr, prop_id)
@@ -60,10 +63,41 @@ struct Property
         else
             val = MosquittoCwrapper.property_read_nonpair(prop_ptr, prop_id, type)
         end
-    
-        return new( name, val, prop_id, type)
+
+        return new(name, val, prop_id, type)
     end
+
+    function Property(name::String, value::T) where {T}
+        # get type and name of property
+        msg_nr, prop, type = MosquittoCwrapper.string_to_property_info(name)
+
+        # possible fallback to User-Property
+        if msg_nr != MosquittoCwrapper.MOSQ_ERR_SUCCESS
+            !(T == String) && error("MQTT property $name can only be converted to a user property. A user property requires String as value type, got $T")
+            prop = MosquittoCwrapper.MQTT_PROP_USER_PROPERTY
+            type = MosquittoCwrapper.MQTT_PROP_TYPE_STRING_PAIR
+            value_uint8 = MosquittoCwrapper.getbytes(value)
+            return new(name, value_uint8, prop, type)
+        end
+
+        # All other types except User-Property. Check value against correct property type
+        Trequired = MosquittoCwrapper.get_julia_type(type)
+        @assert T == Trequired "Expected type $Trequired for property $prop."
+        value_uint8 = MosquittoCwrapper.getbytes(value)
+
+        return new(name, value_uint8, prop, type)
+    end
+
+end # struct Property
+
+function show(io::IO, ::MIME"text/plain", property::Property)
+    show(io, property.name)
+    print(io, ':')
+    show(io, Symbol(property.type))
 end
+
+show(io::IO, property::Property) = show(io, MIME("text/plain"), property)
+
 
 @inline function isvalidproperty(prop_ptr::Ptr{CmosquittoProperty})
     _, success = MosquittoCwrapper.property_identifier(prop_ptr)
@@ -96,6 +130,7 @@ end
 
 """
     add_property!(proplist::PropertyList, name::String, value::T) where {T}
+    add_property!(proplist::PropertyList, property::Property)
 
 Add a property to an existing property list.
 See help of `MosquittoCwrapper.mqtt5_property` and `MosquittoCwrapper.property_identifier_to_string` for valid inputs.
@@ -107,9 +142,9 @@ function add_property!(proplist::PropertyList, name::String, value::T) where {T}
 
     # case of no mqtt5_property found => try MQTT_PROP_USER_PROPERTY
     if msg_nr != MosquittoCwrapper.MOSQ_ERR_SUCCESS
-        !(T==String) && error("MQTT property $name can only be added as type user property. A user property requires String as value type, got $T")
+        !(T == String) && error("MQTT property $name can only be added as type user property. A user property requires String as value type, got $T")
         prop = MosquittoCwrapper.MQTT_PROP_USER_PROPERTY
-        MosquittoCwrapper.property_add_string_pair(proplist.mosq_prop, prop, (name=>value))
+        MosquittoCwrapper.property_add_string_pair(proplist.mosq_prop, prop, (name => value))
         return proplist
     end
 
@@ -141,6 +176,43 @@ function add_property!(proplist::PropertyList, name::String, value::T) where {T}
     end
 
     return proplist
+end
+
+
+function add_property!(proplist::PropertyList, property::Property)
+    # assumes fit of property.prop and property.type which
+    # should be given by the constructor of `Property`.
+    # no further check is performed
+
+    if property.type == MosquittoCwrapper.MQTT_PROP_TYPE_BYTE
+        @assert(length(property.value) == 1, "MQTT_PROP_TYPE_BYTE requires a single byte as value, got $(length(property.value)) bytes")
+        MosquittoCwrapper.property_add_byte(proplist.mosq_prop, property.prop, property.value[1])
+
+    elseif property.type == MosquittoCwrapper.MQTT_PROP_TYPE_INT16
+        @assert(length(property.value) == 2, "MQTT_PROP_TYPE_INT16 requires two bytes as value, got $(length(property.value)) bytes")
+        MosquittoCwrapper.property_add_int16(proplist.mosq_prop, property.prop, reinterpret(UInt16, property.value)[1])
+
+    elseif property.type == MosquittoCwrapper.MQTT_PROP_TYPE_INT32
+        @assert(length(property.value) == 4, "MQTT_PROP_TYPE_INT32 requires four bytes as value, got $(length(property.value)) bytes")
+        MosquittoCwrapper.property_add_int32(proplist.mosq_prop, property.prop, reinterpret(UInt32, property.value)[1])
+
+    elseif property.type == MosquittoCwrapper.MQTT_PROP_TYPE_VARINT
+        @assert(length(property.value) == 4, "MQTT_PROP_TYPE_VARINT requires four bytes as value, got $(length(property.value)) bytes")
+        MosquittoCwrapper.property_add_varint(proplist.mosq_prop, property.prop, reinterpret(UInt32, property.value)[1])
+
+    elseif property.type == MosquittoCwrapper.MQTT_PROP_TYPE_BINARY
+        MosquittoCwrapper.property_add_binary(proplist.mosq_prop, property.prop, property.value)
+
+    elseif property.type == MosquittoCwrapper.MQTT_PROP_TYPE_STRING
+        # TODO could do here without copying
+        value = String(copy(property.value))
+        MosquittoCwrapper.property_add_string(proplist.mosq_prop, property.prop, value)
+
+    else # MosquittoCwrapper.MQTT_PROP_TYPE_STRING_PAIR
+        value = String(copy(property.value))
+        MosquittoCwrapper.property_add_string_pair(proplist.mosq_prop, property.prop, property.name => value)
+    end
+    return nothing
 end